rh264: compare the weighted prediction denominators as unsigned
luma_log2_denom and chroma_log2_denom are bounded to 0..7 by 7.4.3.2 and
the parser checks for that, but the check is a signed upper-bound test on
an int holding the result of rh264_ue. An over-long code saturates to
0xFFFFFFFF, which lands in the int as -1, is not greater than 7, and so
passes: the unit-weight defaults are then built with 1 << -1, and the
value goes on to serve as the shift in the weighting itself (8.4.2.3).

Compare as unsigned, matching the guards on coded_block_pattern and
sub_mb_type.

This came out of auditing every Exp-Golomb read whose value later reaches
an index or a shift, rather than waiting for a fuzzer to arrive at it -
the same defect had already been fixed three times at other sites, and
the weight table is only parsed when weighted prediction is signalled, so
generic mutation reaches it rarely. Confirmed with mutation confined to
the slice headers of the weighted prediction streams, which reproduces
the negative shift at both the luma and chroma sites and is clean after
the change.

All forty-seven regression streams remain byte-identical to the reference
decoder on the Annex-B and AVCC paths, and 700 further mutation
iterations run clean under ASan+UBSan+LSan.

+++ libretro-common/formats/h264/rh264.c
@@ -491,8 +491,12 @@ static int rh264_parse_slice_header_adv(rh264_bits *b,int nal_unit_type,int nal_
       sh->chroma_log2_denom=rh264_ue(b);
       /* 7.4.3.2 bounds both denominators to 0..7; a corrupt stream can
        * signal anything, and the unit-weight defaults below shift by
-       * the value read. */
-      if(sh->luma_log2_denom>7||sh->chroma_log2_denom>7)
+       * the value read.  Compare as unsigned: rh264_ue saturates to
+       * 0xFFFFFFFF on an over-long code, which lands in these ints as
+       * -1 and slips past a signed upper-bound test, leaving the shifts
+       * below - and the ones in the weighting itself - with a negative
+       * exponent. */
+      if((unsigned)sh->luma_log2_denom>7||(unsigned)sh->chroma_log2_denom>7)
          return 0;
       for(i=0;i<32;i++){
          sh->wp_lw[i]=(int16_t)(1<<sh->luma_log2_denom); sh->wp_lo[i]=0;
